Moved ProductGlossary to glossary.py
